+++ data-open-weather-remote/src/main/java/com/example/open_weather/forecast/Daily.kt
@@ -29,7 +29,7 @@ data class Daily(
     @SerializedName("clouds")
     val clouds: Int?,
     @SerializedName("pop")
-    val pop: Int?,
+    val pop: Double?,
     @SerializedName("rain")
     val rain: Double?,
     @SerializedName("uvi")
